fix settings UI
